apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
redis[asyncio]==5.1.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
boto3==1.35.30
stripe==10.12.0
pyotp==2.9.0
